+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Tracking consent dialog
 - Ability to share multiple files (#5)
 - Settings screen & ability to tweak appearance settings for better app performance (#97)
+- Spanish localization
 
 ### Changed
 - Main screen UI changes (#82)